Extract KtCompiler to detekt-parser preparing a minimal test utils module
