+++ detekt-parser/build.gradle.kts
@@ -1,5 +1,7 @@
 dependencies {
     api(kotlin("compiler-embeddable"))
+    implementation(project(":detekt-psi-utils"))
+    testImplementation(project(":detekt-test-utils"))
 }
 
 tasks.withType<Test> {
