+++ detekt-parser/src/test/resources/cases/Default.kt
@@ -1,0 +1,7 @@
+package cases
+
+/**
+ * A comment
+ */
+@Suppress("Unused")
+class Default
