+++ detekt-psi-utils/build.gradle.kts
@@ -1,4 +1,3 @@
 dependencies {
     implementation(kotlin("compiler-embeddable"))
 }
-
