+++ detekt-test-utils/build.gradle.kts
@@ -1,0 +1,3 @@
+dependencies {
+    implementation(project(":detekt-parser"))
+}
