+++ detekt-test/build.gradle.kts
@@ -1,6 +1,7 @@
 dependencies {
     api(kotlin("stdlib-jdk8"))
     api(project(":detekt-core"))
+    api(project(":detekt-test-utils"))
     implementation(kotlin("script-runtime"))
     implementation(kotlin("script-util"))
     implementation(kotlin("scripting-compiler-embeddable"))
