+++ detekt-test/src/main/kotlin/io/gitlab/arturbosch/detekt/test/Resources.kt
@@ -3,13 +3,6 @@ package io.gitlab.arturbosch.detekt.test
 import io.gitlab.arturbosch.detekt.api.internal.YamlConfig
 import java.net.URI
 
-internal object Resources
-
-fun resource(name: String): URI {
-    val explicitName = if (name.startsWith("/")) name else "/$name"
-    val resource = Resources::class.java.getResource(explicitName)
-    requireNotNull(resource) { "Make sure the resource '$name' exists!" }
-    return resource.toURI()
-}
+fun resource(name: String): URI = io.github.detekt.test.utils.resource(name)
 
 fun yamlConfig(name: String) = YamlConfig.loadResource(resource(name).toURL())
